requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
requests==2.26.0
soupsieve==2.2.1
beautifulsoup4==4.9.3
pitchfork-api==0.0.4
